refactored directories, added js and styles directories, modified import paths
