+++ src/js/components/PerformanceChartComponent.js
