+++ src/js/components/VideoRecordingComponent.js
