+++ src/js/pages/Home.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./Home.css";
+import "../../styles/pages/Home.css";
 
 function Home() {
   return (
