+++ src/js/pages/Register.js
